Move App into a module
